Refactor codebase for Voice Enhancement
